fix(restructure): cross-platform lockstep audit — 6 verified divergences (both engines) (#58)
A 2-agent deep audit of the restructure pipeline + IPC/DB contract (each
finding independently verified against code) surfaced 6 real divergences. The
learn-from-corrections code itself audited clean; these are in surrounding /
pre-existing code.

- macOS folderClassifications + per-move tier (F-C1-004 parity): macOS computed
  the Keep/Tidy/Junk classification on the ALREADY-STRIPPED proposals without the
  semantic-claim exemption, so the "Keep" tile silently undercounted the folders
  actually left alone and tier badges diverged from Windows. Now computed in
  proposeAll on the FULL pre-strip set with the exemption (new pure
  folderTiersAndCounts helper + PlanResult), mirroring the Windows
  handle_plan_restructure loop. Pinned by a new exemption test.
- category_counts nondeterminism (Rust): sorted by count only over a HashMap, so
  equal-count categories ordered randomly run-to-run on Windows AND diverged from
  macOS. Added the count-desc-then-category-asc tie-break to match macOS + be
  deterministic.
- filename_tokens length unit (Swift): counted grapheme clusters vs Rust's
  Unicode scalars, diverging on NFD multi-scalar tokens. Swift now counts
  unicodeScalars to match the canonical Rust engine.
- distinctive-term idf precision (Rust): computed ln in f32 (libm-dependent,
  can differ Windows<->macOS by ~1 ULP and flip a group folder name). Now f64
  ln narrowed to f32, matching macOS + portable.
- FileDoneEvent.skippedStages (Swift IPC): non-optional with synthesized Codable
  would throw keyNotFound on a frame that omits it (schema marks it optional,
  Rust elides when empty). Added a tolerant custom init(from:) defaulting to [].
- destFolder root guard (Swift): returned "/" for a root-level destination where
  Rust returns None; aligned (defensive, unreachable in production).
- migrations.rs header comment: listed up to v17, omitted v18 (cosmetic).

Both engines green: Rust 349 lib tests + clippy --all-targets -D warnings clean;
macOS build + 233 tests.

+++ platforms/windows/src/engine/src/pipeline/restructure_semantic.rs
@@ -738,7 +738,12 @@ fn distinctive_terms<'a>(
         .map(|(t, c)| {
             let tf = c as f32 / size;
             let df = *global_freq.get(t).unwrap_or(&1) as f32;
-            (t, tf * (total / df).ln().max(0.0))
+            // Compute the idf's ln in f64 then narrow — f32 `logf` is libm-dependent
+            // (can differ Windows↔macOS by ~1 ULP), which could flip the tie order of
+            // two near-equal-score terms and pick a DIFFERENT group folder name across
+            // platforms. f64 `ln` is consistently correctly-rounded, matching the macOS
+            // engine's `Float(log(Double(...)))`. (audit — lockstep)
+            (t, tf * (((total / df) as f64).ln().max(0.0) as f32))
         })
         .collect();
     // Score desc, then name for determinism; drop zero-score (ubiquitous) tags.
